test: 로그인 테스트 케이스 추가 및 테스트 생명주기 변경 (메서드 -> 클래스)
